requirements.txt
mcp>=1.0.0
fastmcp>=0.2.0
cookidoo-api>=0.8.0
python-dotenv>=1.0.0
pydantic>=2.0.0
